refacted leaderboard to another controller
